+++ quantification-entity/src/main/java/com/huobi/quantification/entity/QuanDepthFutureDetail.java
@@ -5,36 +5,36 @@ import java.util.Date;
 
 public class QuanDepthFutureDetail {
     /**
-     * @mbg.generated 2018-06-28 15:27:19
+     * @mbg.generated 2018-07-02 14:32:37
      */
     private Long id;
 
     /**
      * 深度ID
-     * @mbg.generated 2018-06-28 15:27:19
+     * @mbg.generated 2018-07-02 14:32:37
      */
     private Long depthFutureId;
 
     /**
-     * @mbg.generated 2018-06-28 15:27:19
+     * @mbg.generated 2018-07-02 14:32:37
      */
     private Integer detailType;
 
     /**
      * 价格
-     * @mbg.generated 2018-06-28 15:27:19
+     * @mbg.generated 2018-07-02 14:32:37
      */
     private BigDecimal detailPrice;
 
     /**
      * 数量
-     * @mbg.generated 2018-06-28 15:27:19
+     * @mbg.generated 2018-07-02 14:32:37
      */
-    private Double detailAmount;
+    private BigDecimal detailAmount;
 
     /**
      * 更新时间
-     * @mbg.generated 2018-06-28 15:27:19
+     * @mbg.generated 2018-07-02 14:32:37
      */
     private Date dateUpdate;
 
@@ -70,11 +70,11 @@ public class QuanDepthFutureDetail {
         this.detailPrice = detailPrice;
     }
 
-    public Double getDetailAmount() {
+    public BigDecimal getDetailAmount() {
         return detailAmount;
     }
 
-    public void setDetailAmount(Double detailAmount) {
+    public void setDetailAmount(BigDecimal detailAmount) {
         this.detailAmount = detailAmount;
     }
